+++ cli/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Features
+* The enclave for `module exec` will now be named after the module image and the time it was ran
+
 ### Fixes
 * Fixed a bug where `service logs` that was successful would really fail