+++ cli/docs/changelog.md
@@ -1,5 +1,13 @@
 # TBD
 
+# 0.7.2
+### Features
+* The enclave for `module exec` will now be named after the module image and the time it was ran
+* Allow users running `module exec` to manually specify the ID of the enclave that will be created
+
+### Fixes
+* Fixed a bug where `service logs` that was successful would really fail
+
 # 0.7.1
 ### Fixes
 * Attempt to fix CLI artifact publishing